Partial Skeleton
Added the database using a list type structure to implement it (including persisting it with json).
More routines need to be added to the database.py file.
Slightly modified bin.py to use numeric IDs instead of names (since that is what a database will ultimately need for the primary key). Also, added a record file to store past full weights and times. Changed latitude and longitude to use shifted decimal place integer.